# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: gameManageKit API
  version: 1.2.0
  description: 多游戏、HTTP-only、MySQL-only 的账号门户服务。
servers:
  - url: https://portal.example.com
    description: Public
  - url: http://game-manage-kit.internal:2571
    description: Internal
...
paths:
  /v1/admin/auth/login:
    post:
      operationId: adminLogin
      tags: [Admin]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '204':
          description: 登录成功；会话仅通过 HttpOnly Cookie 返回
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/admin/auth/session:
    get:
      operationId: getAdminSession
      tags: [Admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: 当前管理员及实时游戏权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
    delete:
      operationId: adminLogout
      tags: [Admin]
      security:
        - AdminSession: []
      responses:
        '204':
          description: 会话已撤销并清除 Cookie
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/admin/games:
    get:
      operationId: listAdminGames
      tags: [Admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: 返回全部游戏项目及其配置状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameProjectListResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createAdminGame
      description: 创建草稿游戏项目；必须使用管理员 Cookie 会话，并校验请求 Origin。
      tags: [Admin]
      security:
        - AdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameProjectRequest'
      responses:
        '201':
          description: 游戏项目已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/admin/games/{gameId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
    patch:
      operationId: updateAdminGame
      description: 按 revision 编辑游戏项目；gameId 不可修改，且必须校验请求 Origin。
      tags: [Admin]
      security:
        - AdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameProjectRequest'
      responses:
        '200':
          description: 游戏项目已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/admin/games/{gameId}/servers:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      operationId: listAdminGameServers
      description: 返回指定游戏的全部区服，包括尚未开放给客户端的区服。
      tags: [Admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: 游戏区服列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedGameServerListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createAdminGameServer
      description: 为指定游戏新增区服；必须使用管理员 Cookie 会话，并校验请求 Origin。
      tags: [Admin]
      security:
        - AdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameServerRequest'
      responses:
        '201':
          description: 游戏区服已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedGameServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/admin/games/{gameId}/servers/{serverId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
      - $ref: '#/components/parameters/ServerId'
    patch:
      operationId: updateAdminGameServer
      description: 按 revision 编辑区服；gameId 与 serverId 不可修改，且必须校验请求 Origin。
      tags: [Admin]
      security:
        - AdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameServerRequest'
      responses:
        '200':
          description: 游戏区服已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedGameServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
    ServiceId:
      type: apiKey
      in: header
      name: x-service-id
    ServiceSecret:
      type: apiKey
      in: header
      name: x-service-secret
    AdminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
    OperatorId:
      type: apiKey
      in: header
      name: x-operator-id
    AdminSession:
      type: apiKey
      in: cookie
      name: __Host-gmk_admin_session
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/GameId'
...
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 65535
  responses:
    BadRequest:
      description: 入参不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    AdminUnauthorized:
      description: 管理员凭证或会话无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 禁止访问；可能为账号封禁、游戏停用、调用方无游戏权限或区服停用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源、游戏或区服不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 操作冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: 请求过频
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Internal:
      description: 内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unavailable:
      description: 微信上游不可用，或游戏处于维护状态
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GameId:
      type: string
      minLength: 2
      maxLength: 32
      pattern: '^[a-z][a-z0-9-]{1,31}$'
    GameStatus:
      type: string
      enum: [enabled, maintenance, disabled]
    GameConfigurationState:
      type: string
      enum: [draft, configured]
    ErrorCode:
      type: string
      enum:
        - INVALID_PAYLOAD
        - AUTH_REQUIRED
        - ACCOUNT_BANNED
        - NOT_FOUND
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - SERVICE_AUTH_REQUIRED
        - SERVICE_FORBIDDEN
        - OPERATION_CONFLICT
        - INTERNAL
        - GAME_NOT_FOUND
        - GAME_DISABLED
        - GAME_ACCESS_DENIED
        - SERVER_NOT_FOUND
        - SERVER_DISABLED
        - ADMIN_AUTH_REQUIRED
        - ORIGIN_FORBIDDEN
        - GAME_PROJECT_CONFLICT
        - GAME_SERVER_CONFLICT
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        requestId:
          type: string
...
    AdminLoginRequest:
      type: object
      additionalProperties: false
      required: [operatorId, password]
      properties:
        operatorId:
          type: string
          maxLength: 64
        password:
          type: string
          maxLength: 1024
    AdminOperator:
      type: object
      additionalProperties: false
      required: [operatorId, displayName]
      properties:
        operatorId:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[a-z][a-z0-9_.-]{2,63}$'
        displayName:
          type: string
          minLength: 1
          maxLength: 128
    AdminGameAccess:
      type: object
      additionalProperties: false
      required: [gameId, name, status, canOperateAccounts]
      properties:
        gameId:
          $ref: '#/components/schemas/GameId'
        name:
          type: string
          minLength: 1
          maxLength: 128
        status:
          $ref: '#/components/schemas/GameStatus'
        canOperateAccounts:
          type: boolean
    AdminSessionResponse:
      type: object
      additionalProperties: false
      required: [operator, games, canManageGames, expiresAt]
      properties:
        operator:
          $ref: '#/components/schemas/AdminOperator'
        games:
          type: array
          items:
            $ref: '#/components/schemas/AdminGameAccess'
        canManageGames:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    GameProject:
      type: object
      additionalProperties: false
      required:
        - gameId
        - name
        - description
        - status
        - configurationState
        - clientVisible
        - sortOrder
        - revision
        - createdAt
        - updatedAt
      properties:
        gameId:
          $ref: '#/components/schemas/GameId'
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/GameStatus'
        configurationState:
          $ref: '#/components/schemas/GameConfigurationState'
        clientVisible:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 65535
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GameProjectListResponse:
      type: object
      additionalProperties: false
      required: [games]
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/GameProject'
    CreateGameProjectRequest:
      type: object
      additionalProperties: false
      required: [gameId, name, description]
      properties:
        gameId:
          $ref: '#/components/schemas/GameId'
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 500
    UpdateGameProjectRequest:
      type: object
      additionalProperties: false
      required: [name, description, status, clientVisible, sortOrder, revision]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/GameStatus'
        clientVisible:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 65535
        revision:
          type: integer
          minimum: 1
    ManagedGameServer:
      type: object
      additionalProperties: false
      required:
        - gameId
        - serverId
        - name
        - tag
        - status
        - openTime
        - gameHttpUrl
        - gameWsUrl
        - isOpen
        - sortOrder
        - revision
        - createdAt
        - updatedAt
      properties:
        gameId:
          $ref: '#/components/schemas/GameId'
        serverId:
          type: integer
          minimum: 0
          maximum: 65535
        name:
          type: string
          minLength: 1
          maxLength: 64
        tag:
          type: string
          enum: [normal, new, full, maintenance]
        status:
          type: string
          enum: [smooth, busy, maintenance]
        openTime:
          type: integer
          minimum: 0
        gameHttpUrl:
          type: string
          format: uri
          maxLength: 2048
        gameWsUrl:
          type: string
          format: uri
          maxLength: 2048
        isOpen:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 65535
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ManagedGameServerListResponse:
      type: object
      additionalProperties: false
      required: [servers]
      properties:
        servers:
          type: array
          maxItems: 65536
          items:
            $ref: '#/components/schemas/ManagedGameServer'
    CreateGameServerRequest:
      type: object
      additionalProperties: false
      required:
        - serverId
        - name
        - tag
        - status
        - openTime
        - gameHttpUrl
        - gameWsUrl
        - isOpen
        - sortOrder
      properties:
        serverId:
          type: integer
          minimum: 0
          maximum: 65535
        name:
          type: string
          minLength: 1
          maxLength: 64
        tag:
          type: string
          enum: [normal, new, full, maintenance]
        status:
          type: string
          enum: [smooth, busy, maintenance]
        openTime:
          type: integer
          minimum: 0
        gameHttpUrl:
          type: string
          format: uri
          maxLength: 2048
        gameWsUrl:
          type: string
          format: uri
          maxLength: 2048
        isOpen:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 65535
    UpdateGameServerRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - tag
        - status
        - openTime
        - gameHttpUrl
        - gameWsUrl
        - isOpen
        - sortOrder
        - revision
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        tag:
          type: string
          enum: [normal, new, full, maintenance]
        status:
          type: string
          enum: [smooth, busy, maintenance]
        openTime:
          type: integer
          minimum: 0
        gameHttpUrl:
          type: string
          format: uri
          maxLength: 2048
        gameWsUrl:
          type: string
          format: uri
          maxLength: 2048
        isOpen:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 65535
        revision:
          type: integer
          minimum: 1
...